% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary2.R
\name{country_incidence_summary}
\alias{country_incidence_summary}
\title{Summarize measles incidence rate per 1000000 of the population for a country}
\usage{
country_incidence_summary(country_iso3)
}
\arguments{
\item{country_iso3}{iso3 of country to be analyzed}
}
\value{
a data frame containing the country's urban classification
the mean and standard deviation of measles
incidence rates for a selected country
}
\description{
Summarize measles incidence rate per 1000000 of the population for a country
}
