% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2.R
\name{country_plot}
\alias{country_plot}
\title{A graph of each countries measles cases per month}
\usage{
country_plot(country_name)
}
\arguments{
\item{country_name}{A character vector of country names}
}
\value{
A graph
}
\description{
A graph of each countries measles cases per month
}
