% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load measles data from a CSV URL}
\usage{
load_data()
}
\value{
A tibble with the loaded dataset.
}
\description{
This function downloads and reads the measles data sets used in the analysis.
}
