% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{load_urban_data}
\alias{load_urban_data}
\title{A lollipop graph of each countries mean incidence rate.}
\usage{
load_urban_data()
}
\value{
A graph
}
\description{
A lollipop graph of each countries mean incidence rate.
}
