% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mean_measles_by_country}
\alias{mean_measles_by_country}
\title{A lollipop graph of each countries mean incidence rate.}
\usage{
mean_measles_by_country(urban_data, countries)
}
\arguments{
\item{countries}{A character vector of country names}
}
\value{
A graph
}
\description{
A lollipop graph of each countries mean incidence rate.
}
