% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical-analysis.R
\name{measles_model_setup}
\alias{measles_model_setup}
\title{Sets up data to be used to create model table}
\usage{
measles_model_setup(countries = NULL)
}
\arguments{
\item{countries}{Optional character vector of countries to include.}
}
\value{
A dataframe with measles proportion per month and urban population rates for specified countries
}
\description{
Sets up data to be used to create model table
}
