% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wrangling.R
\name{prepare_mean_measles_data}
\alias{prepare_mean_measles_data}
\title{Merges and filters data for mean measles incidence plot}
\usage{
prepare_mean_measles_data(urban_data, countries)
}
\arguments{
\item{urban_data}{A data frame from urban_data()}

\item{countries}{A character vector of country names}
}
\value{
A data frame
}
\description{
Merges and filters data for mean measles incidence plot
}
