% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2.R
\name{summ_sus_measles_per_month}
\alias{summ_sus_measles_per_month}
\title{cleans data}
\usage{
summ_sus_measles_per_month(country_name)
}
\arguments{
\item{country_name}{A character vector of country names}
}
\value{
a tibble
}
\description{
cleans data
}
