% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-urban.R
\name{summary_urban}
\alias{summary_urban}
\title{Summarize measles incidence rates by urban population classification}
\usage{
summary_urban(country)
}
\arguments{
\item{country_iso3}{countries}
}
\value{
A tibble
}
\description{
Summarize measles incidence rates by urban population classification
}
