% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{urban_data}
\alias{urban_data}
\title{Joins the urban data set with urban classifications and the merged urban data set}
\usage{
urban_data()
}
\value{
a dataframe
}
\description{
Joins the urban data set with urban classifications and the merged urban data set
}
