% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wrangling.R
\name{urban_type_incident}
\alias{urban_type_incident}
\title{Selects countries from urban data set and gives urban classification and mean measles incident rate}
\usage{
urban_type_incident(df, country_iso3)
}
\arguments{
\item{df}{dataframe with urban populations}

\item{country_iso3}{iso3 of countries to be analyzed}
}
\value{
a data frame
}
\description{
Selects countries from urban data set and gives urban classification and mean measles incident rate
}
