% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-urban.R
\name{mean_measles_per_month_for_country}
\alias{mean_measles_per_month_for_country}
\title{A table for the selected country that includes mean and sd measles incidence.}
\usage{
mean_measles_per_month_for_country(country_name)
}
\arguments{
\item{country}{A country name}
}
\value{
A data frame with the country, mean incidence, and standard deviation of incidence.
}
\description{
A table for the selected country that includes mean and sd measles incidence.
}
