% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{summ_sus_measles_per_month}
\alias{summ_sus_measles_per_month}
\title{A plot of suspected measles cases per month for a selected country}
\usage{
summ_sus_measles_per_month(country)
}
\value{
A graph
}
\description{
A plot of suspected measles cases per month for a selected country
}
